// packages/dataviz/src/types.d.ts
import {
  BarChartSettings,
  LineChartSettings,
  PieChartSettings,
  MetricChartSettings,
  FunnelChartSettings,
} from '@keen.io/charts';
import { CardSettings } from '@keen.io/ui-core';
import { TextSettings, LegendSettings } from '@keen.io/widgets';
import { Widgets } from './render-widget';
export declare type VisualizerWidgetSettings = {
  title?: Partial<TextSettings>;
  subtitle?: Partial<TextSettings>;
  legend?: Partial<LegendSettings>;
  card?: Partial<CardSettings>;
};
export declare type ComponentSettings =
  | BarChartSettings
  | LineChartSettings
  | PieChartSettings
  | MetricChartSettings
  | FunnelChartSettings
  | {};
export declare type Options = {
  container: HTMLElement | string;
  type: Widgets;
  widget?: Partial<VisualizerWidgetSettings>;
  settings?: ComponentSettings;
};